// shared/data/src/mocks/students-slider/students-slider.interface.ts
export interface StudentsSlider {
  fullName: string
  age: string
  profession: string
  description: string
  rotateCard?: number
  firstIcon?: object
  verticalPositionFirstIcon?: number
  horizontalPositionFirstIcon?: number
  secondIcon?: object
  verticalPositionSecondIcon?: number
  horizontalPositionSecondIcon?: number
}